Adapt tests for new framework version.

+++ node/src/constant.rs
@@ -1,5 +1,5 @@
 /// Unique service ID.
-pub const BLOCKCHAIN_SERVICE_ID: u16 = 128;
+pub const BLOCKCHAIN_SERVICE_ID: u32 = 128;
 /// Name of the service.
 pub const BLOCKCHAIN_SERVICE_NAME: &str = "cryptoelection_blockchain_service";
 /// Time oracle instance name.
